+++ src/io/net/basic_serializer.h
@@ -68,14 +68,14 @@ public:
 
     std::string encode_counts(data_struct::Stream_Block* in_stream_block);
 
-	data_struct::Stream_Block* decode_counts(char* message, int message_len);
+    data_struct::Stream_Block* decode_counts(char* message, size_t message_len);
 
 protected:
 	template <typename T>
 	void _convert_var_to_bytes(std::string * str, char* bytes_temp, T variable, size_t size)
 	{
 		memcpy(bytes_temp, (char*)(&variable), size);
-		for (int i = 0; i < size; i++)
+        for (size_t i = 0; i < size; i++)
 			*str += bytes_temp[i];
 	}
